Roles permissions

See merge request BI-TWA/B211/team-dfk!23

+++ backend/reservations/views/reservation.py
@@ -1,3 +1,4 @@
+from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse, Http404
 from django.shortcuts import redirect, render
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
@@ -18,15 +19,19 @@ class ReservationViewSet(viewsets.ModelViewSet):
 
 class ReservationTemplateView(ListView):
     @staticmethod
+    @login_required()
     def reservation_get_view(request, reservation_id):
         reservation = ReservationService.find_by_id(reservation_id)
+        if reservation not in ReservationService.find_reservations_for_person(request.user):
+            redirect("/login")
         template = loader.get_template("reservations/detail.html")
         return HttpResponse(template.render({"reservation": reservation}, request))
 
     @staticmethod
+    @login_required()
     def reservations_get_view(request):
         page = request.GET.get("page", 1)
-        paginator = Paginator(ReservationService.find_all(), DEFAULT_PAGE_SIZE)
+        paginator = Paginator(ReservationService.find_reservations_for_person(request.user), DEFAULT_PAGE_SIZE)
         try:
             reservations = paginator.page(page)
         except PageNotAnInteger:
@@ -37,34 +42,25 @@ class ReservationTemplateView(ListView):
         return render(request, "reservations/publiclist.html", {"reservations": reservations})
 
     @staticmethod
+    @login_required()
     def reservation_delete_view(request, reservation_id):
+        reservation = ReservationService.find_by_id(reservation_id)
+        if reservation not in ReservationService.find_reservations_for_person(request.user):
+            redirect("/login")
+
         template = loader.get_template("reservations/publiclist.html")
         if not ReservationService.delete(reservation_id):
             return HttpResponse(template.render({"errors": ["Failed to delete reservation"]}, request))
         return redirect("/reservations/")
 
     @staticmethod
+    @login_required
     def reservation_create_view(request):
-        form = ReservationForm(request.POST or None)
-        if form.is_valid():
-            timestamp = datetime.now()
-            reservation_status = ReservationStatus()
-            reservation_status.author = form.cleaned_data.get("author")
-            reservation_status.modified = timestamp
-            reservation_status.note = form.cleaned_data.get("note")
-            reservation_status.save()
-
-            reservation = Reservation()
-            reservation.author = form.cleaned_data.get("author")
-            reservation.dt_from = form.cleaned_data.get("dt_from")
-            reservation.dt_to = form.cleaned_data.get("dt_to")
-            reservation.dt_to = form.cleaned_data.get("dt_to")
-            reservation.dt_created = timestamp
-            reservation.room = form.cleaned_data.get("room")
-            reservation.save()
-
-            reservation.attendees.set(form.cleaned_data.get("attendees"))
-            reservation.reservation_status.add(reservation_status)
-            return redirect("/reservations/")
+        form = ReservationForm(request, request.POST or None)
         template = loader.get_template("reservations/create.html")
+
+        if form.is_valid():
+            if not ReservationService.save(form.cleaned_data, request.user):
+                return HttpResponse(template.render({"errors": ["Something went wrong"], "form": form}, request))
+            return redirect("/")
         return HttpResponse(template.render({"form": form}, request))
